v2/requirements.txt
atproto>=0.0.40
python-dotenv>=1.0.0
feedparser>=6.0.10
requests>=2.28.0
Pillow>=10.0.0
beautifulsoup4>=4.12.0
pytz>=2024.1
tzlocal>=5.0.0
Jinja2>=3.1.0
